Move `Application` and rename httpApplication pkg
`Application` was moved to the httpserver package to remove
the cyclic package relationship between httpapplication and
httpserver. HttpApplication was renamed clientApplication to make
its purpose clearer.
